backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-multipart==0.0.20
pydantic==2.11.7
google-generativeai==0.3.2
sentence-transformers==2.5.1
python-dotenv==1.1.0
html2text==2025.4.15
Pillow==11.2.1